requirements.txt
lightrag==0.1.0b6
  backoff==2.2.1
  Jinja2==3.1.6
    MarkupSafe==3.0.2
  jsonlines==4.0.0
    attrs==25.3.0
  nest-asyncio==1.6.0
  numpy==1.26.4
  python-dotenv==1.1.0
  PyYAML==6.0.2
  tiktoken==0.7.0
    regex==2024.11.6
    requests==2.32.3
      charset-normalizer==3.4.2
      idna==3.10
      urllib3==2.4.0
      certifi==2025.4.26
  tqdm==4.67.1
networkx
spacy
sentence-transformers
faiss-cpu
pubmed_parser
